+++ server/apps/immich/src/controllers/index.ts
@@ -1,1 +1,2 @@
+export * from './api-key.controller';
 export * from './user.controller';
