+++ server/libs/domain/src/api-key/api-key.service.ts
@@ -2,11 +2,11 @@ import { UserEntity } from '@app/infra';
 import { BadRequestException, Inject, Injectable, UnauthorizedException } from '@nestjs/common';
 import { compareSync, hash } from 'bcrypt';
 import { randomBytes } from 'node:crypto';
-import { AuthUserDto } from '../../decorators/auth-user.decorator';
+import { AuthUserDto } from '../auth';
 import { IKeyRepository } from './api-key.repository';
 import { APIKeyCreateDto } from './dto/api-key-create.dto';
-import { APIKeyCreateResponseDto } from './repsonse-dto/api-key-create-response.dto';
-import { APIKeyResponseDto, mapKey } from './repsonse-dto/api-key-response.dto';
+import { APIKeyCreateResponseDto } from './response-dto/api-key-create-response.dto';
+import { APIKeyResponseDto, mapKey } from './response-dto/api-key-response.dto';
 
 @Injectable()
 export class APIKeyService {
@@ -58,14 +58,22 @@ export class APIKeyService {
     return keys.map(mapKey);
   }
 
-  async validate(token: string): Promise<UserEntity> {
+  async validate(token: string): Promise<AuthUserDto> {
     const [_id, key] = Buffer.from(token, 'base64').toString('utf8').split(':');
     const id = Number(_id);
 
     if (id && key) {
       const entity = await this.repository.getKey(id);
       if (entity?.user && entity?.key && compareSync(key, entity.key)) {
-        return entity.user as UserEntity;
+        const user = entity.user as UserEntity;
+
+        return {
+          id: user.id,
+          email: user.email,
+          isAdmin: user.isAdmin,
+          isPublicUser: false,
+          isAllowUpload: true,
+        };
       }
     }
