+++ server/libs/domain/src/api-key/dto/index.ts
@@ -1,0 +1,2 @@
+export * from './api-key-create.dto';
+export * from './api-key-update.dto';
