+++ server/libs/domain/src/api-key/response-dto/index.ts
@@ -1,0 +1,2 @@
+export * from './api-key-create-response.dto';
+export * from './api-key-response.dto';
