+++ server/libs/domain/src/index.ts
@@ -1,3 +1,4 @@
+export * from './api-key';
 export * from './auth';
 export * from './domain.module';
 export * from './user';
