+++ server/libs/infra/src/db/repository/api-key.repository.ts
@@ -1,22 +1,8 @@
-import { APIKeyEntity } from '@app/infra';
+import { IKeyRepository } from '@app/domain';
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-
-export const IKeyRepository = 'IKeyRepository';
-
-export interface IKeyRepository {
-  create(dto: Partial<APIKeyEntity>): Promise<APIKeyEntity>;
-  update(userId: string, id: number, dto: Partial<APIKeyEntity>): Promise<APIKeyEntity>;
-  delete(userId: string, id: number): Promise<void>;
-  /**
-   * Includes the hashed `key` for verification
-   * @param id
-   */
-  getKey(id: number): Promise<APIKeyEntity | null>;
-  getById(userId: string, id: number): Promise<APIKeyEntity | null>;
-  getByUserId(userId: string): Promise<APIKeyEntity[]>;
-}
+import { APIKeyEntity } from '../entities';
 
 @Injectable()
 export class APIKeyRepository implements IKeyRepository {
