+++ server/libs/infra/src/db/repository/index.ts
@@ -1,1 +1,2 @@
+export * from './api-key.repository';
 export * from './user.repository';
