+++ server/libs/infra/src/infra.module.ts
@@ -1,11 +1,13 @@
+import { IKeyRepository, IUserRepository } from '@app/domain';
 import { databaseConfig, UserEntity } from '@app/infra';
-import { IUserRepository } from '@app/domain';
 import { Global, Module, Provider } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { UserRepository } from './db';
+import { APIKeyEntity, UserRepository } from './db';
+import { APIKeyRepository } from './db/repository';
 
 const providers: Provider[] = [
   //
+  { provide: IKeyRepository, useClass: APIKeyRepository },
   { provide: IUserRepository, useClass: UserRepository },
 ];
 
@@ -14,7 +16,7 @@ const providers: Provider[] = [
   imports: [
     //
     TypeOrmModule.forRoot(databaseConfig),
-    TypeOrmModule.forFeature([UserEntity]),
+    TypeOrmModule.forFeature([APIKeyEntity, UserEntity]),
   ],
   providers: [...providers],
   exports: [...providers],
